feat: add bandersnatch ring root calculation using ffi
Added bandersnatch ring root calculation using the foreign function
interface (ffi) to the crypto module.  The ring root is composed of the
Bandersnatch keys of each validator in the next epoch. The calculation
is used in the transition function of the safrole module to update the
post_state.gamma_z field.