feat: safrole full and dynamic ring size in ring vrf
Added support for dynamic ring size in ring vrf. Previously, the ring
size was hardcoded to 6 or 1023. Now, the ring size can be passed as an
argument to the `ring_context` function, allowing for more flexibility
in the ring vrf implementation. Additionally, the `PCS_PARAMS` and
`RING_CONTEXT_CACHE` are now initialized with the dynamic ring size.

The changes are reflected in the `safrole` module, where the `Params`
struct now includes a `validators_count` field, and the `transition`
function takes the ring size into account when verifying ticket
envelopes.